Update App.java
Fixed NPEs for cases of CAS not being found and empty rows in sheet before final row reached. 
Added check that CAS is actually matched for the goodscents site (fragLookup) since that will also return results that contain the query within the CAS. 
Changed to a direct URL lookup for fragLookup instead of using htmlUnitDriver since it wasn't needed for that site. 
Added exit with error message for case of the sheet not existing